refactor(xtask): extract accuracy types into parser/accuracy.rs to fix LOC-cap test
The `test_update_status_is_split_into_subsystem_modules` test in mod.rs
asserts each subsystem module under `update_status/` stays at or below
400 lines. `parser.rs` had grown to 524 lines — 124 lines over the cap —
causing the test to fail silently on master (xtask unit tests are excluded
from `cargo test --workspace --lib`).

Extract all parser-accuracy concerns into a new sibling module
`parser/accuracy.rs`:
- `ParserAccuracyArtifactSummary` and its three nested structs
- `read_parser_accuracy_artifact` (reads `target/metrics/parser_accuracy.json`)
- `parser_accuracy_rows`, `parser_accuracy_family_summary`,
  `parser_accuracy_metric_summary`

Move `PARSER_STATUS_MARKER_NAMES` (test-only constant) from `parser.rs`
to `parser/tests.rs` where it's exclusively used.

Add explicit imports to `parser/tests.rs` for the accuracy types (no longer
reachable via `use super::*` now that they live in the child module) and for
the `token` module (no longer imported by `parser.rs` in cfg(test) context).

Result: `parser.rs` drops from 524 → 375 lines; all five files checked by
the test are now under 400 lines. The LOC-cap test passes.

No behaviour change. No changes to production crates.

Fixes #7642
